Extract the editor output code from the update-output-repos script, because they need to run on different triggers
